Small theme change to menu. Menu will be replaced.
I gotta say, I'm not a fan of how this menu system looks. And with
how big this menu system is, I'm not sure I want to keep plugging
away at theming it and making it look nice.

I'm considering replacing it with a much, much smaller menu
system with some static sprite buttons and a text entry field
for entering puzzle ID numbers. There don't need to be any
sub menus or