RELEASE 1.1.0
This update adds in "real time" map modifications with no change
to the map library itself. Each map now has a prefixed 2-digit number
that corresponds to each modification done on the map. The modifications
include:

    Map rotation
    Map flipping

So, while the true number of maps hasn't changed, they are flipped
and rotated when loaded, increasing the puzzle variations from
~50,000 to nearly 400,000.

The save files have been updated to track the currently open map
with the modifications done.